+++ src/bench/fdm/benchmarkFDM.cpp
@@ -14,7 +14,7 @@ namespace{
 struct CallParameters{
   int Nelements;
   int Nq_e;
-  int wordSize;
+  size_t wordSize;
   bool useRAS;
   bool overlap;
   std::string suffix;
@@ -42,7 +42,7 @@ std::map<CallParameters, occa::kernel> cachedResults;
 template <typename T>
 occa::kernel benchmarkFDM(int Nelements,
                           int Nq_e,
-                          int wordSize,
+                          size_t wordSize,
                           bool useRAS,
                           bool overlap,
                           int verbosity,
@@ -300,7 +300,7 @@ occa::kernel benchmarkFDM(int Nelements,
 
 template occa::kernel benchmarkFDM<int>(int Nelements,
                                         int Nq_e,
-                                        int wordSize,
+                                        size_t wordSize,
                                         bool useRAS,
                                         bool overlap,
                                         int verbosity,
@@ -310,7 +310,7 @@ template occa::kernel benchmarkFDM<int>(int Nelements,
 
 template occa::kernel benchmarkFDM<double>(int Nelements,
                                            int Nq_e,
-                                           int wordSize,
+                                           size_t wordSize,
                                            bool useRAS,
                                            bool overlap,
                                            int verbosity,
